+++ lib/facebook/messenger/incoming/pass_thread_control.rb
@@ -1,12 +1,17 @@
 module Facebook
   module Messenger
     module Incoming
-      # The PolicyEnforcement class represents an incoming webhook response from
-      # Facebook when they are notifying your app of a policy violation
+      # The PassThreadControl class represents an incoming Facebook Messenger
+      # pass thread control event.
       #
-      # https://developers.facebook.com/docs/messenger-platform/reference/webhook-events/messaging_policy_enforcement
+      # @see https://developers.facebook.com/docs/messenger-platform/handover-protocol/pass-thread-control
+      # @see https://developers.facebook.com/docs/messenger-platform/reference/handover-protocol/pass-thread-control
       class PassThreadControl
         include Facebook::Messenger::Incoming::Common
+
+        def new_owner_app_id
+          @messaging['pass_thread_control']['new_owner_app_id']
+        end
 
         def metadata
           @messaging['pass_thread_control']['metadata']
@@ -15,11 +20,6 @@ module Facebook
         def sender_id
           @messaging['sender']['id']
         end
-
-        def new_owner_app_id
-          @messaging['pass_thread_control']['new_owner_app_id']
-        end
-
       end
     end
   end