+++ lib/facebook/messenger/subscriptions.rb
@@ -9,7 +9,7 @@ module Facebook
     module Subscriptions
       include HTTParty
 
-      base_uri 'https://graph.facebook.com/v2.9/me'
+      base_uri 'https://graph.facebook.com/v3.2/me'
 
       format :json
 
@@ -27,9 +27,10 @@ module Facebook
       #
       # @return [Boolean] TRUE
       #
-      def subscribe(access_token:)
+      def subscribe(access_token:, subscribed_fields: [])
         response = post '/subscribed_apps', query: {
-          access_token: access_token
+          access_token: access_token,
+          subscribed_fields: subscribed_fields
         }
 
         raise_errors(response)