+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
 
+## [1.3.0] - 2019-04-16
+
+### Added
+- Support for PassThreadControl
+
+## [1.2.0] - 2019-04-13
+
+### Added
+- Upgraded to Graph API 3.2 for everything
+- Support for appsecret_proof
+
+## [1.1.2] - 2018-12-06
+
+### Added
+- Support for Graph API 3.2 in `Subscriptions`.
+
 ## [1.1.1] - 2018-07-10
 
 ### Fixed